+++ src/vue/modules/create-opportunity/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="create-opportunity">
+  <div class="create-opportunity" v-if="isInitialized">
     <form-stepper
       :steps="STEPS"
       :current-step.sync="currentStep"
@@ -14,7 +14,7 @@
       >
         <div class="create-opportunity__cards">
           <opportunity-card
-            v-for="item in formTypes"
+            v-for="item in FORM_TYPES"
             :key="item.value"
             :is-select="form.information.formType.value === item.value"
             :title="item.label | globalize"
@@ -97,10 +97,10 @@
             <input-field
               white-autofill
               type="number"
-              v-model="form.information.annualReturn"
-              @blur="touchField('form.information.annualReturn')"
-              id="hard-cap"
-              name="create-sale-hard-cap"
+              v-model="form.saleInformation.annualReturn"
+              @blur="touchField('form.saleInformation.annualReturn')"
+              id="anual-return"
+              name="create-sale-anual-return"
               :label="
                 form.information.formType.value ===
                   ASSET_SUBTYPE.bond ?
@@ -108,10 +108,31 @@
                     'create-opportunity.expected-revenue' | globalize
               "
               :error-message="getFieldErrorMessage(
-                'form.information.annualReturn',
+                'form.saleInformation.annualReturn',
                 {
-                  from:'0',
-                  to:maxAmount
+                  from:MIN_ALLOWED_PERCENT,
+                  to:MAX_ALLOWED_PERCENT
+                }
+              )"
+              :disabled="formMixin.isDisabled"
+            />
+          </div>
+        </div>
+        <div class="app__form-row">
+          <div class="app__form-field">
+            <input-field
+              white-autofill
+              type="number"
+              v-model="form.information.maxIssuanceAmount"
+              @blur="touchField('form.information.maxIssuanceAmount')"
+              id="max-issuance-amount"
+              name="create-sale-max-issuance-amount"
+              :label="'create-opportunity.max-issuance-amount' | globalize"
+              :error-message="getFieldErrorMessage(
+                'form.information.maxIssuanceAmount',
+                {
+                  from: minAmount,
+                  to: maxAmount
                 }
               )"
               :disabled="formMixin.isDisabled"
@@ -149,6 +170,51 @@
         class="app__form"
         @submit.prevent="nextStep('form.saleInformation')"
       >
+        <div class="app__form-row">
+          {{ 'create-opportunity.accept-investments-in' | globalize }}
+        </div>
+        <template v-if="form.information.formType.value !== ASSET_SUBTYPE.bond">
+          <div
+            class="app__form-row"
+            v-for="item in baseAssets"
+            :key="item.code"
+          >
+            <div class="app__form-field">
+              <tick-field
+                :name="`create-sale-tick-${item.code}`"
+                v-model="form.saleInformation.quoteAssets"
+                :disabled="formMixin.isDisabled"
+                :cb-value="item.code"
+              >
+                {{ item.nameAndCode }}
+              </tick-field>
+            </div>
+          </div>
+        </template>
+        <template v-else>
+          <div
+            class="app__form-row"
+            v-for="item in baseAssets"
+            :key="item.code"
+          >
+            <div class="app__form-field">
+              <radio-field
+                :key="`radio-field-${item.code}`"
+                name="create-sale-radio-asset-code"
+                v-model="form.saleInformation.quoteAssets"
+                :active="form.saleInformation.quoteAssets"
+                :cb-value="item.code"
+                :disabled="formMixin.isDisabled"
+              >
+                {{ item.nameAndCode }}
+              </radio-field>
+            </div>
+          </div>
+        </template>
+        <div class="create-opportunity__error-text">
+          {{ getFieldErrorMessage('form.saleInformation.quoteAssets') }}
+        </div>
+
         <div class="app__form-row">
           <div class="app__form-field">
             <date-field
@@ -205,13 +271,12 @@
               id="soft-cap"
               name="create-sale-soft-cap"
               :label="'create-opportunity.soft-cap' | globalize({
-                asset: statsQuoteAsset.code
+                asset: form.saleInformation.defaultQuoteAsset
               })"
               :error-message="getFieldErrorMessage(
                 'form.saleInformation.softCap',
                 {
-                  from:minAmount,
-                  to:maxAmount
+                  hardCap: form.saleInformation.hardCap || '0'
                 }
               )"
               :disabled="formMixin.isDisabled"
@@ -228,13 +293,12 @@
               id="hard-cap"
               name="create-sale-hard-cap"
               :label="'create-opportunity.hard-cap' | globalize({
-                asset: statsQuoteAsset.code
+                asset: form.saleInformation.defaultQuoteAsset
               })"
               :error-message="getFieldErrorMessage(
                 'form.saleInformation.hardCap',
                 {
-                  from:form.saleInformation.softCap,
-                  to:maxAmount
+                  softCap: form.saleInformation.softCap || '0'
                 }
               )"
               :disabled="formMixin.isDisabled"
@@ -270,50 +334,6 @@
               {{ 'create-opportunity.transferable-lbl' | globalize }}
             </tick-field>
           </div>
-        </div>
-        <div class="app__form-row">
-          {{ 'create-opportunity.accept-investments-in' | globalize }}
-        </div>
-        <template v-if="form.information.formType.value !== ASSET_SUBTYPE.bond">
-          <div
-            class="app__form-row"
-            v-for="item in baseAssets"
-            :key="item.code"
-          >
-            <div class="app__form-field">
-              <tick-field
-                :name="`create-sale-tick-${item.code}`"
-                v-model="form.saleInformation.quoteAssets"
-                :disabled="formMixin.isDisabled"
-                :cb-value="item.code"
-              >
-                {{ item.nameAndCode }}
-              </tick-field>
-            </div>
-          </div>
-        </template>
-        <template v-else>
-          <div
-            class="app__form-row"
-            v-for="item in baseAssets"
-            :key="item.code"
-          >
-            <div class="app__form-field">
-              <radio-field
-                :key="`radio-field-${item.code}`"
-                :name="`create-sale-radio-asset-code`"
-                v-model="form.saleInformation.quoteAssets"
-                :active="form.saleInformation.quoteAssets"
-                :cb-value="item.code"
-                :disabled="formMixin.isDisabled"
-              >
-                {{ item.nameAndCode }}
-              </radio-field>
-            </div>
-          </div>
-        </template>
-        <div class="create-opportunity__error-text">
-          {{ getFieldErrorMessage('form.saleInformation.quoteAssets') }}
         </div>
         <div class="app__form-actions">
           <button
@@ -395,11 +415,11 @@
         </div>
       </form>
     </form-stepper>
-    <loader
-      v-else
-      message-id="create-opportunity.loading-msg"
-    />
   </div>
+  <loader
+    v-else
+    message-id="create-opportunity.loading-msg"
+  />
 </template>
 
 <script>
@@ -413,19 +433,21 @@ import moment from 'moment'
 import { DocumentUploader } from '@/js/helpers/document-uploader'
 import { DOCUMENT_TYPES } from '@/js/const/document-types.const'
 import { Wallet, base, ASSET_POLICIES, SALE_TYPES } from '@tokend/js-sdk'
-import { AssetRecord } from './wrappers/asset.record'
 import { api, initApi } from './_api'
 import { BLOB_TYPES } from '@/js/const/blob-types.const'
 import { ASSET_SUBTYPE, ASSET_SUBTYPE_IMG_URL } from '@/js/const/asset-subtypes.const'
 
 import { DateUtil } from '@/js/utils'
-import { mapActions } from 'vuex'
+import { mapActions, mapGetters } from 'vuex'
+import { MathUtil } from '@/js/utils/math.util'
 import { types } from './store/types'
 import {
   required,
   amountRange,
   maxLength,
   requiredAtLeastOne,
+  softCapMoreThanHardCap,
+  hardCapLessThanSoftCap,
   minDate,
   maxDate,
 } from '@validators'
@@ -450,6 +472,8 @@ const STEPS = {
 }
 const CODE_MAX_LENGTH = 16
 const NAME_MAX_LENGTH = 255
+const MIN_ALLOWED_PERCENT = 0
+const MAX_ALLOWED_PERCENT = 100
 const DESCRIPTION_MAX_LENGTH = 255
 const DEFAULT_SALE_TYPE = '0'
 const ASSET_CREATION_REQUEST_ID = '0'
@@ -461,6 +485,19 @@ const EMPTY_DOCUMENT = {
 const EVENTS = {
   close: 'close',
 }
+const FORM_TYPES = [
+  {
+    label: 'create-opportunity.bond-creation',
+    value: ASSET_SUBTYPE.bond,
+    description: 'create-opportunity.bond-description',
+  },
+  {
+    label: 'create-opportunity.property-purchase',
+    value: ASSET_SUBTYPE.share,
+    description: 'create-opportunity.revenue-description',
+  },
+]
+
 export default {
   name: 'create-opportunity',
   components: {
@@ -505,6 +542,7 @@ export default {
         information: {
           name: '',
           code: '',
+          maxIssuanceAmount: '',
           formType: {},
           terms: null,
         },
@@ -518,6 +556,7 @@ export default {
           quoteAssets: [],
           assetType: '',
           policies: [],
+          defaultQuoteAsset: '',
         },
         shortBlurb: {
           saleLogo: null,
@@ -526,16 +565,20 @@ export default {
           description: '',
         },
       },
+      isInitialized: false,
       isSubmitting: false,
       currentStep: 1,
       STEPS,
       kvAssetTypeKycRequired: '',
+      FORM_TYPES,
       CODE_MAX_LENGTH,
       ASSET_SUBTYPE,
       DOCUMENT_TYPES,
       ASSET_POLICIES,
       NAME_MAX_LENGTH,
       DESCRIPTION_MAX_LENGTH,
+      MIN_ALLOWED_PERCENT,
+      MAX_ALLOWED_PERCENT,
     }
   },
   validations () {
@@ -568,10 +611,12 @@ export default {
             maxLength: maxLength(CODE_MAX_LENGTH),
           },
           maturityDate: maturityDate,
-          annualReturn: {
+          maxIssuanceAmount: {
             required,
-            amountRange: amountRange('0',
-              this.maxAmount),
+            amountRange: amountRange(
+              this.minAmount,
+              this.maxAmount
+            ),
           },
         },
         saleInformation: {
@@ -579,18 +624,30 @@ export default {
           endTime: endTime,
           softCap: {
             required,
-            amountRange: amountRange(this.minAmount, this.maxAmount),
+            softCapMoreThanHardCap: softCapMoreThanHardCap(
+              this.minAmount,
+              this.form.saleInformation.hardCap
+            ),
           },
           hardCap: {
             required,
-            amountRange: amountRange(this.form.saleInformation.softCap,
-              this.maxAmount),
+            hardCapLessThanSoftCap: hardCapLessThanSoftCap(
+              this.form.saleInformation.softCap,
+              this.maxAmount
+            ),
           },
           quoteAssets: {
             requiredAtLeastOne,
           },
           assetType: {
             required,
+          },
+          annualReturn: {
+            required,
+            amountRange: amountRange(
+              '0',
+              '100'
+            ),
           },
         },
         shortBlurb: {
@@ -606,20 +663,12 @@ export default {
     }
   },
   computed: {
-    formTypes () {
-      return [
-        {
-          label: 'create-opportunity.bond-creation',
-          value: ASSET_SUBTYPE.bond,
-          description: 'create-opportunity.bond-description',
-        },
-        {
-          label: 'create-opportunity.property-purchase',
-          value: ASSET_SUBTYPE.share,
-          description: 'create-opportunity.revenue-description',
-        },
-      ]
-    },
+    ...mapGetters('create-opportunity', {
+      assets: types.assets,
+      pairs: types.pairs,
+      baseAssets: types.baseAssets,
+      statsQuoteAsset: types.statsQuoteAsset,
+    }),
     assetTypes () {
       return [
         {
@@ -632,19 +681,37 @@ export default {
         },
       ]
     },
-    baseAssets () {
-      return this.assets.filter(item => item.isBaseAsset)
+    isBond () {
+      return this.form.information.formType.value === ASSET_SUBTYPE.bond
     },
-    statsQuoteAsset () {
-      return this.assets.find(item => item.isStatsQuoteAsset)
+    salePriceRatioStatsQuoteAsset () {
+      return MathUtil.divide(
+        this.form.saleInformation.hardCap,
+        this.form.information.maxIssuanceAmount
+      )
+    },
+  },
+  watch: {
+    'form.information.formType.value' () {
+      this.assignDefaultAssetSubtype()
+      this.assignDefaultQuoteAsset()
+    },
+    'form.saleInformation.quoteAssets' () {
+      this.assignDefaultQuoteAsset()
+    },
+    'form.saleInformation.defaultQuoteAsset' () {
+      this.form.saleInformation.softCap = ''
+      this.form.saleInformation.hardCap = ''
     },
   },
   async created () {
     initApi(this.wallet, this.config)
-    this.form.information.formType = this.formTypes[0]
-    const response = await this.loadAssets()
-    this.assets = response.map(item => new AssetRecord(item))
+    this.form.information.formType = this.FORM_TYPES[0]
+    await this.loadAssets()
+    await this.loadBaseAssetsPairs()
     this.kvAssetTypeKycRequired = await this.loadKvAssetTypeKycRequired()
+    this.assignDefaultAssetSubtype()
+    this.isInitialized = true
   },
   methods: {
     moment,
@@ -653,6 +720,7 @@ export default {
       loadKvAssetTypeKycRequired: types.LOAD_KV_KYC_REQUIRED,
       getBlobId: types.LOAD_BLOB_ID,
       loadAssets: types.LOAD_ASSETS,
+      loadBaseAssetsPairs: types.LOAD_BASE_ASSETS_PAIRS_BY_STATS_QUOTE_ASSET,
     }),
     nextStep (formStep) {
       if (this.isFormValid(formStep)) {
@@ -698,7 +766,7 @@ export default {
       const operation = {
         requestID: this.isUpdateMode ? this.request.id : '0',
         baseAsset: this.form.information.code,
-        defaultQuoteAsset: this.statsQuoteAsset.code,
+        defaultQuoteAsset: this.form.saleInformation.defaultQuoteAsset,
         startTime: DateUtil.toTimestamp(this.form.saleInformation.startTime),
         endTime: DateUtil.toTimestamp(this.form.saleInformation.endTime),
         softCap: this.form.saleInformation.softCap,
@@ -709,11 +777,15 @@ export default {
           description: saleDescriptionBlobId,
           logo: this.form.shortBlurb.saleLogo.getDetailsForSave(),
           youtube_video_id: this.form.shortBlurb.youtubeId,
+          annualReturn: this.form.saleInformation.annualReturn,
         },
-        requiredBaseAssetForHardCap: this.form.saleInformation.hardCap,
+        requiredBaseAssetForHardCap: this.form.information.maxIssuanceAmount,
         quoteAssets: this.formatSaleQuoteAssets(),
         saleEnumType: SALE_TYPES.basicSale,
         saleType: DEFAULT_SALE_TYPE,
+      }
+      if (this.isBond) {
+        operation.defaultQuoteAsset = this.form.saleInformation.quoteAssets
       }
       return base.SaleRequestBuilder.createSaleCreationRequest(operation)
     },
@@ -729,22 +801,34 @@ export default {
         assetType: this.form.saleInformation.assetType.value,
         preissuedAssetSigner: this.accountId,
         trailingDigitsCount: this.decimalPints,
-        initialPreissuedAmount: this.form.saleInformation.hardCap,
-        maxIssuanceAmount: this.form.saleInformation.hardCap,
+        initialPreissuedAmount: this.form.information.maxIssuanceAmount,
+        maxIssuanceAmount: this.form.information.maxIssuanceAmount,
         policies: this.form.saleInformation.policies
           .reduce((a, b) => (a | b), 0),
         creatorDetails: {
           name: this.form.information.name,
           logo: EMPTY_DOCUMENT,
           terms: terms ? terms.getDetailsForSave() : EMPTY_DOCUMENT,
-          annualReturn: this.form.information.annualReturn,
+          annualReturn: this.form.saleInformation.annualReturn,
           subtype: this.form.information.formType.value,
         },
       }
-      if (this.form.information.formType.value === ASSET_SUBTYPE.bond) {
+      if (this.isBond) {
         operation.creatorDetails.maturityDate = DateUtil
           .toMs(this.form.information.maturityDate)
         operation.creatorDetails.logoUrl = ASSET_SUBTYPE_IMG_URL.bondLogo
+        operation.creatorDetails.investmentToken = {
+          asset: this.form.saleInformation.quoteAssets,
+          price: this.salePriceRatioStatsQuoteAsset,
+        }
+
+        operation.creatorDetails.redeemPrice = MathUtil.add(
+          MathUtil.percentOfValue(
+            this.salePriceRatioStatsQuoteAsset,
+            this.form.saleInformation.annualReturn
+          ),
+          this.salePriceRatioStatsQuoteAsset
+        )
       } else {
         operation.creatorDetails.logoUrl = ASSET_SUBTYPE_IMG_URL.shareLogo
       }
@@ -765,59 +849,103 @@ export default {
       }
     },
     formatSaleQuoteAssets () {
-      if (Array.isArray(this.form.saleInformation.quoteAssets)) {
-        return this.form.saleInformation.quoteAssets
-          .map((item) => ({
-            asset: item,
-            price: '1',
-          }))
-      } else {
+      if (this.isBond) {
         return [{
           asset: this.form.saleInformation.quoteAssets,
-          price: '1',
+          price: this.salePriceRatioStatsQuoteAsset,
         }]
+      } else {
+        if (this.form.saleInformation.quoteAssets.length > 1) {
+          return this.form.saleInformation.quoteAssets
+            .map(asset => ({
+              asset: asset,
+              price: this.calculatePriceForBaseAsset(asset),
+            }))
+        } else {
+          return [{
+            asset: this.form.saleInformation.quoteAssets[0],
+            price: this.salePriceRatioStatsQuoteAsset,
+          }]
+        }
       }
     },
     selectFormType (formType) {
       this.form.information.maturityDate = ''
       this.form.information.formType = formType
     },
+    assignDefaultAssetSubtype () {
+      if (this.isBond && this.baseAssets.length) {
+        this.form.saleInformation.quoteAssets = this.baseAssets[0].code
+      } else {
+        this.form.saleInformation.quoteAssets = []
+      }
+    },
+    assignDefaultQuoteAsset () {
+      const quoteAssets = this.form.saleInformation.quoteAssets
+      if (!this.isBond) {
+        if (quoteAssets.length > 1) {
+          // eslint-disable-next-line max-len
+          this.form.saleInformation.defaultQuoteAsset = this.statsQuoteAsset.code
+        } else {
+          // eslint-disable-next-line max-len
+          this.form.saleInformation.defaultQuoteAsset = quoteAssets[0] || this.statsQuoteAsset.code
+        }
+      } else {
+        // eslint-disable-next-line max-len
+        this.form.saleInformation.defaultQuoteAsset = this.form.saleInformation.quoteAssets
+      }
+    },
+    calculatePriceForBaseAsset (asset) {
+      if (asset === this.statsQuoteAsset.code) {
+        return this.salePriceRatioStatsQuoteAsset
+      } else {
+        const assetPrice = this.pairs.find(item =>
+          item.baseAsset.id === asset &&
+          item.quoteAsset.id === this.statsQuoteAsset.code
+        ).price || '1'
+
+        return MathUtil.divide(
+          this.salePriceRatioStatsQuoteAsset,
+          assetPrice
+        )
+      }
+    },
   },
 }
 </script>
 
-<style lang="scss" scoped>
-  @import '@/vue/forms/_app-form';
+<style lang="scss">
+@import '@/vue/forms/_app-form';
 
-  .create-opportunity__btn {
-    margin-bottom: 2rem;
-    width: 14.4rem;
-  }
+.create-opportunity__btn {
+  margin-bottom: 2rem;
+  width: 14.4rem;
+}
 
-  .create-opportunity__kyc-required-row {
-    margin-top: 2.1rem;
-  }
+.create-opportunity__kyc-required-row {
+  margin-top: 2.1rem;
+}
 
-  .create-opportunity__pre-issued-asset-signer-wrp {
-    display: flex;
-    align-items: center;
-  }
+.create-opportunity__pre-issued-asset-signer-wrp {
+  display: flex;
+  align-items: center;
+}
 
-  .create-opportunity__insert-account-id-btn {
-    margin-left: .4rem;
-  }
+.create-opportunity__insert-account-id-btn {
+  margin-left: .4rem;
+}
 
-  .create-opportunity__error-text {
-    margin-bottom: 2rem;
-    color: $col-error;
-  }
+.create-opportunity__error-text {
+  margin-bottom: 2rem;
+  color: $col-error;
+}
 
-  .create-opportunity__price {
-    font-size: 1.4rem;
-  }
+.create-opportunity__price {
+  font-size: 1.4rem;
+}
 
-  .create-opportunity__cards {
-    width: 100%;
-    display: flex;
-  }
+.create-opportunity__cards {
+  width: 100%;
+  display: flex;
+}
 </style>