+++ CHANGELOG.md
@@ -35,6 +35,14 @@ for further information about branching and tagging conventions.
 - Application logo URL for module scheme
 - Displaying user avatar in passport
 - Loyalty points merchant & reconcillation schemes
+- Subject info `Dividend for <token-name>` for every payment operation in dividends
+- `All opportunities(All sales)` and `My opportunities(My sales)` pages on the `Opportunities(Funds)` page
+- For sales that were ended show highlighted text like a "Closed [days] ago"
+- Reword create opportunity label "KYC Required" => "Verification required"
+- Asset loading via the loop where was `filter[limit]=100`
+- Correct base/quote price calculating for Opportunities
+- Custom validation rules for sale's "Soft cap" and "Hard cap" values
+- Custom Movements TopBar for REIT scheme
 
 ### Changed
 - Now using new account endpoint for loading the fees
@@ -42,6 +50,7 @@ for further information about branching and tagging conventions.
 - Renamed `DescriptionEditor` -> `MarkdownField`
 - Moved `MarkdownField` to the fields & included it to the form mixin components
 - Use @tokend/js-sdk@1.3.1-x.2
+- Moved TopBar to the module
 
 ### Fixed
 - Sidebar displaying in the Safari browser on the small screens
@@ -51,6 +60,22 @@ for further information about branching and tagging conventions.
 - Move no-data-message translate ids pass to the elements that use component
 - Chart line displaying
 - Loading balances instead of account on withdrawal form
+- Investment in the sale. We could not invest the full amount before, because the amount was calculated not according to the base asset, but by quote
+- Loader displaying in the Opportunity create form
+- Calculated redeem price displaying for old tokens
+- Correct dividends subject
+- Fees displaying on the Movements page
+- "Soft cap" validation in the "Opportunity create" form
+- Asset name displaying
+- Change defaultQuoteAsset to single acceptable asset for the Revenue opportunities
+
+## [1.3.1-rc.0] - 2019-03-20
+
+### Changed
+- Do not throw an error if movement has an unknown effect or operation details
+
+### Fixed
+- Not handling 'manage-asset-pair' operation details
 
 ## [1.3.0] - 2019-03-01